noticket: FIX: YdbTableHint.int64Range() used `int` (which is `INT32` range)
...instead of `long` (which would be proper for `INT64` range)

+++ repository-ydb-v2/src/main/java/tech/ydb/yoj/repository/ydb/client/YdbTableHint.java
@@ -18,7 +18,7 @@ public class YdbTableHint {
     PartitioningSettings partitioningSettings;
     TablePreset tablePreset;
 
-    public static YdbTableHint int64Range(int start, int end) {
+    public static YdbTableHint int64Range(long start, long end) {
         return new YdbTableHint(
                 new PartitioningPolicy()
                         .setAutoPartitioning(AutoPartitioningPolicy.DISABLED)
